size_t

+++ Day01/ex00/libft/ft_strsplit.c
@@ -6,7 +6,7 @@
 /*   By: jblack-b <jblack-b@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/30 19:25:35 by jblack-b          #+#    #+#             */
-/*   Updated: 2018/12/09 19:22:19 by jblack-b         ###   ########.fr       */
+/*   Updated: 2018/12/09 19:23:41 by jblack-b         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -50,7 +50,7 @@ static int		ft_wlen(const char *s, char c)
 
 static int		ft_strsplit_cleanup(char ***dst)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (dst[i])
